Split vector and stack into separate files as should be properly done. Also separated headers and source files into sub-directories for legibility.
